fix: FakeRepository를 사용하기 위해 DI를 활용한 방식으로 repository 이름 변경 HoneyTipRepository -> HoneyTipJpaRepository
