+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/image/HoneyTipImageJpaRepository.java
@@ -1,0 +1,9 @@
+package com.api.ttoklip.domain.honeytip.repository.image;
+
+import com.api.ttoklip.domain.honeytip.domain.HoneyTipImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HoneyTipImageJpaRepository extends JpaRepository<HoneyTipImage, Long> {
+
+    boolean existsByHoneyTipIdAndUrl(final Long honeyTipId, final String url);
+}
