+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/image/HoneyTipImageRepository.java
@@ -1,9 +1,17 @@
 package com.api.ttoklip.domain.honeytip.repository.image;
 
 import com.api.ttoklip.domain.honeytip.domain.HoneyTipImage;
-import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
 
-public interface HoneyTipImageRepository extends JpaRepository<HoneyTipImage, Long>, HoneyTipImageRepositoryCustom {
+public interface HoneyTipImageRepository {
 
-    boolean existsByHoneyTipIdAndUrl(final Long honeyTipId, final String url);
+    void save(HoneyTipImage honeyTipImage);
+
+    boolean existsByHoneyTipIdAndUrl(Long honeyTipId, String url);
+
+    void verifyMemberIsImageOwner(List<Long> imageIds, Long memberId);
+
+    boolean doAllImageIdsExist(List<Long> imageIds);
+
+    void deleteByImageIds(List<Long> imageIds);
 }
