--- src/main/java/com/api/ttoklip/domain/honeytip/repository/image/HoneyTipImageRepositoryCustom.java
@@ -1,11 +1,0 @@
-package com.api.ttoklip.domain.honeytip.repository.image;
-
-import java.util.List;
-public interface HoneyTipImageRepositoryCustom {
-
-    void allImageOwner(List<Long> imageIds, Long memberId);
-
-    boolean doAllImageIdsExist(List<Long> imageIds);
-
-    void deleteByImageIds(List<Long> imageIds);
-}
