+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/like/HoneyTipLikeJpaRepository.java
@@ -2,16 +2,10 @@ package com.api.ttoklip.domain.honeytip.repository.like;
 
 import com.api.ttoklip.domain.honeytip.domain.HoneyTipLike;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HoneyTipLikeRepository {
-
+public interface HoneyTipLikeJpaRepository extends JpaRepository<HoneyTipLike, Long> {
     Optional<HoneyTipLike> findByHoneyTipIdAndMemberId(Long honeyTipId, Long memberId);
 
     boolean existsByHoneyTipIdAndMemberId(Long honeyTipId, Long memberId);
-
-    Long countHoneyTipLikesByHoneyTipId(final Long postId);
-
-    void save(HoneyTipLike honeyTipLike);
-
-    void deleteById(Long id);
 }
