--- src/main/java/com/api/ttoklip/domain/honeytip/repository/like/HoneyTipLikeRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.api.ttoklip.domain.honeytip.repository.like;
-
-public interface HoneyTipLikeRepositoryCustom {
-    Long countHoneyTipLikesByHoneyTipId(final Long postId);
-}
