+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/like/HoneyTipLikeRepositoryImpl.java
@@ -1,24 +1,39 @@
 package com.api.ttoklip.domain.honeytip.repository.like;
 
-import com.api.ttoklip.domain.honeytip.domain.QHoneyTipLike;
-import com.querydsl.core.types.dsl.Wildcard;
-import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.api.ttoklip.domain.honeytip.domain.HoneyTipLike;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
 
+@Repository
 @RequiredArgsConstructor
-public class HoneyTipLikeRepositoryImpl implements HoneyTipLikeRepositoryCustom {
+public class HoneyTipLikeRepositoryImpl implements HoneyTipLikeRepository{
 
-    private final JPAQueryFactory jpaQueryFactory;
+    private final HoneyTipLikeJpaRepository honeyTipLikeJpaRepository;
+    private final HoneyTipLikeQueryRepository honeyTipLikeQueryRepository;
 
-    private final QHoneyTipLike honeyTipLike = QHoneyTipLike.honeyTipLike;
+    @Override
+    public Optional<HoneyTipLike> findByHoneyTipIdAndMemberId(final Long honeyTipId, final Long memberId) {
+        return honeyTipLikeJpaRepository.findByHoneyTipIdAndMemberId(honeyTipId, memberId);
+    }
+
+    @Override
+    public boolean existsByHoneyTipIdAndMemberId(final Long honeyTipId, final Long memberId) {
+        return honeyTipLikeJpaRepository.existsByHoneyTipIdAndMemberId(honeyTipId, memberId);
+    }
 
     @Override
     public Long countHoneyTipLikesByHoneyTipId(final Long postId) {
-        return jpaQueryFactory
-                .select(Wildcard.count)
-                .from(honeyTipLike)
-                .where(honeyTipLike.honeyTip.id.eq(postId))
-                .fetchOne();
+        return honeyTipLikeQueryRepository.countHoneyTipLikesByHoneyTipId(postId);
     }
 
+    @Override
+    public void save(final HoneyTipLike honeyTipLike) {
+        honeyTipLikeJpaRepository.save(honeyTipLike);
+    }
+
+    @Override
+    public void deleteById(final Long id) {
+        honeyTipLikeJpaRepository.deleteById(id);
+    }
 }
