+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/post/HoneyTipJpaRepository.java
@@ -1,0 +1,7 @@
+package com.api.ttoklip.domain.honeytip.repository.post;
+
+import com.api.ttoklip.domain.honeytip.domain.HoneyTip;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HoneyTipJpaRepository extends JpaRepository<HoneyTip, Long> {
+}
