+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/post/HoneyTipRepository.java
@@ -1,8 +1,29 @@
 package com.api.ttoklip.domain.honeytip.repository.post;
 
+import com.api.ttoklip.domain.common.Category;
 import com.api.ttoklip.domain.honeytip.domain.HoneyTip;
-import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
-public interface HoneyTipRepository extends JpaRepository<HoneyTip, Long>, HoneyTipRepositoryCustom {
+public interface HoneyTipRepository {
+    HoneyTip findByIdActivated(final Long honeyTipId);
 
+    HoneyTip findHoneyTipWithDetails(final Long postId);
+
+    Page<HoneyTip> matchCategoryPaging(Category category, Pageable pageable);
+
+    List<HoneyTip> findRecent3();
+
+    List<HoneyTip> findHouseworkTips();
+
+    List<HoneyTip> findRecipeTips();
+
+    List<HoneyTip> findSafeLivingTips();
+
+    List<HoneyTip> findWelfarePolicyTips();
+
+    List<HoneyTip> getPopularityTop5();
+
+    void save(HoneyTip honeyTip);
 }
