+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/scrap/HoneyTipScrapJpaRepository.java
@@ -2,15 +2,12 @@ package com.api.ttoklip.domain.honeytip.repository.scrap;
 
 import com.api.ttoklip.domain.honeytip.domain.HoneyTipScrap;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HoneyTipScrapRepository {
+public interface HoneyTipScrapJpaRepository extends JpaRepository<HoneyTipScrap, Long> {
+
     Optional<HoneyTipScrap> findByHoneyTipIdAndMemberId(Long honeyTipId, Long memberId);
 
     boolean existsByHoneyTipIdAndMemberId(Long honeyTipId, Long memberId);
-
-    Long countHoneyTipScrapsByHoneyTipId(final Long postId);
-
-    void deleteById(Long id);
-
-    void save(HoneyTipScrap honeyTipScrap);
 }
+
