--- src/main/java/com/api/ttoklip/domain/honeytip/repository/scrap/HoneyTipScrapRepositoryCustom.java
@@ -1,6 +1,0 @@
-package com.api.ttoklip.domain.honeytip.repository.scrap;
-
-public interface HoneyTipScrapRepositoryCustom {
-
-    Long countHoneyTipScrapsByHoneyTipId(final Long honeyTipId);
-}
