+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/url/HoneyTipUrlJpaRepository.java
@@ -2,13 +2,8 @@ package com.api.ttoklip.domain.honeytip.repository.url;
 
 import com.api.ttoklip.domain.honeytip.domain.HoneyTipUrl;
 import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HoneyTipUrlRepository {
-
+public interface HoneyTipUrlJpaRepository extends JpaRepository<HoneyTipUrl, Long> {
     List<HoneyTipUrl> findByHoneyTipId(Long honeyTipId);
-    void deleteAllByIds(final List<Long> ids);
-
-    void save(HoneyTipUrl honeyTipUrl);
-
-    void saveAll(List<HoneyTipUrl> honeyTipUrlsToSave);
 }
