--- src/main/java/com/api/ttoklip/domain/honeytip/repository/url/HoneyTipUrlRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.api.ttoklip.domain.honeytip.repository.url;
-
-import java.util.List;
-
-public interface HoneyTipUrlRepositoryCustom {
-    void deleteAllByIds(final List<Long> ids);
-}
