+++ src/main/java/com/api/ttoklip/domain/honeytip/repository/url/HoneyTipUrlRepositoryImpl.java
@@ -1,22 +1,35 @@
 package com.api.ttoklip.domain.honeytip.repository.url;
 
-import com.api.ttoklip.domain.honeytip.domain.QHoneyTipUrl;
-import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.api.ttoklip.domain.honeytip.domain.HoneyTipUrl;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
 
+@Repository
 @RequiredArgsConstructor
-public class HoneyTipUrlRepositoryImpl implements HoneyTipUrlRepositoryCustom {
+public class HoneyTipUrlRepositoryImpl implements HoneyTipUrlRepository {
 
-    private final JPAQueryFactory queryFactory;
+    private final HoneyTipUrlJpaRepository honeyTipUrlJpaRepository;
+    private final HoneyTipUrlQueryRepository honeyTipUrlQueryRepository;
 
-    private final QHoneyTipUrl honeyTipUrl = QHoneyTipUrl.honeyTipUrl;
+
+    @Override
+    public List<HoneyTipUrl> findByHoneyTipId(final Long honeyTipId) {
+        return honeyTipUrlJpaRepository.findByHoneyTipId(honeyTipId);
+    }
 
     @Override
     public void deleteAllByIds(final List<Long> ids) {
-        queryFactory
-                .delete(honeyTipUrl)
-                .where(honeyTipUrl.id.in(ids))
-                .execute();
+        honeyTipUrlQueryRepository.deleteAllByIds(ids);
+    }
+
+    @Override
+    public void save(final HoneyTipUrl honeyTipUrl) {
+        honeyTipUrlJpaRepository.save(honeyTipUrl);
+    }
+
+    @Override
+    public void saveAll(final List<HoneyTipUrl> honeyTipUrlsToSave) {
+        honeyTipUrlJpaRepository.saveAll(honeyTipUrlsToSave);
     }
 }
